refactor: #4 CompantDetailPage 컴포넌트 단위로 분할
