--- src/components/Spinner.tsx
@@ -1,5 +1,0 @@
-export default function Spinner() {
-  return (
-    <div className="animate-spin rounded-full h-8 w-8 border-t-2 border-blue-500"></div>
-  );
-}
